refactor(community): update folder structure and rename LabCard components
