+++ apps/community/src/mocks/labs.ts
